Move grouping logic out of subscription item.
