+++ src/tree/v2/ResourceGroupsTreeDataProvider.ts
@@ -1,15 +1,16 @@
 import { AzureExtensionApiProvider } from '@microsoft/vscode-azext-utils/api';
 import * as vscode from 'vscode';
 import { ApplicationResourceProviderManager } from '../../api/v2/providers/ApplicationResourceProviderManager';
-import { ext } from '../../extensionVariables';
 import { localize } from '../../utils/localize';
+import { ApplicationResourceGroupingManager } from './ApplicationResourceGroupingManager';
 import { AzureAccountExtensionApi } from './azure-account.api';
 import { GenericItem } from './GenericItem';
 import { ResourceGroupResourceBase } from './ResourceGroupResourceBase';
 import { SubscriptionItem } from './SubscriptionItem';
 
 export class ResourceGroupsTreeDataProvider extends vscode.Disposable implements vscode.TreeDataProvider<ResourceGroupResourceBase> {
-    private readonly configSubscription: vscode.Disposable;
+    private readonly resourceGroupingManager = new ApplicationResourceGroupingManager();
+    private readonly groupingChangeSubscription: vscode.Disposable;
     private readonly onDidChangeTreeDataEmitter = new vscode.EventEmitter<void | ResourceGroupResourceBase | null | undefined>();
 
     private api: AzureAccountExtensionApi | undefined;
@@ -19,21 +20,18 @@ export class ResourceGroupsTreeDataProvider extends vscode.Disposable implements
     constructor(private readonly resourceProviderManager: ApplicationResourceProviderManager) {
         super(
             () => {
-                this.configSubscription.dispose();
+                this.groupingChangeSubscription.dispose();
                 this.filtersSubscription?.dispose();
                 this.statusSubscription?.dispose();
+
+                this.resourceGroupingManager.dispose();
             });
 
         // TODO: This really belongs on the subscription item, but that then involves disposing of them during refresh,
         //       and I'm not sure of the mechanics of that.  Ideally grouping mode changes shouldn't require new network calls,
         //       as we're just rearranging known items; we might try caching resource items and only calling getTreeItem() on
         //       branch providers during the tree refresh that results from this (rather than getChildren() again).
-        this.configSubscription = vscode.workspace.onDidChangeConfiguration(
-            e => {
-                if (e.affectsConfiguration(`${ext.prefix}.groupBy`)) {
-                    this.onDidChangeTreeDataEmitter.fire();
-                }
-            });
+        this.groupingChangeSubscription = this.resourceGroupingManager.onDidChangeGrouping(() => this.onDidChangeTreeDataEmitter.fire());
     }
 
     onDidChangeTreeData: vscode.Event<void | ResourceGroupResourceBase | null | undefined> = this.onDidChangeTreeDataEmitter.event;
@@ -53,7 +51,7 @@ export class ResourceGroupsTreeDataProvider extends vscode.Disposable implements
                     if (api.filters.length === 0) {
                         return [new GenericItem(localize('noSubscriptions', 'Select Subscriptions...'), { commandId: 'azure-account.selectSubscriptions' })]
                     } else {
-                        return api.filters.map(subscription => new SubscriptionItem(subscription, this.resourceProviderManager));
+                        return api.filters.map(subscription => new SubscriptionItem(subscription, this.resourceGroupingManager, this.resourceProviderManager));
                     }
                 } else if (api.status === 'LoggedOut') {
                     return [
